Made it a bit more reliable

+++ src/components/EditorMenu.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import PromptBox from './PromptBox';
 
 type EditorMenuProps = {
   /**
@@ -6,9 +7,18 @@ type EditorMenuProps = {
    */
   onDelete?: () => void;
   /**
+   * The string a user must type in order to successfully trigger the delete.
+   */
+  deleteConfirmation: string;
+  /**
    * Edit file metadata & details callback.
    */
   onEdit?: () => void;
+}
+
+type EditorMenuState = {
+  confirmDelete: boolean;
+  deleteComplete: boolean;
 }
 
 const buttonStyles: React.CSSProperties = {
@@ -24,13 +34,27 @@ const divStyles: React.CSSProperties = {
   display: 'flex',
   position: 'absolute',
   bottom: '1rem',
-  gap: '1rem'
+  gap: '1rem',
+  marginLeft: '1rem'
 }
 
-export default class EditorMenu extends React.Component<EditorMenuProps> {
+export default class EditorMenu extends React.Component<EditorMenuProps, EditorMenuState> {
 
-  private handleDelete = () => {
-    this.props.onDelete && this.props.onDelete();
+  state = {
+    confirmDelete: false,
+    deleteComplete: false
+  }
+
+  private beginDelete = (): void => {
+    this.setState({
+      confirmDelete: true
+    })
+  }
+
+  private handleDelete = (confirmation: string) => {
+    if (confirmation === this.props.deleteConfirmation) {
+      this.props.onDelete && this.props.onDelete();
+    }
   }
 
   private handleEdit = () => {
@@ -48,9 +72,19 @@ export default class EditorMenu extends React.Component<EditorMenuProps> {
         }
         {
           this.props.onDelete &&
-          <button style={buttonStyles} onClick={this.handleDelete}>
+          <button style={buttonStyles} onClick={this.beginDelete}>
             <i style={iconStyles} className="material-icons">delete</i>
           </button>
+        }
+        {
+          this.state.confirmDelete &&
+          <PromptBox 
+            message={
+              `Are you sure you want to delete this document? This action cannot be undone. To delete, type "${this.props.deleteConfirmation}"`}
+            onSubmit={this.handleDelete}
+            onCancel={() => {this.setState({confirmDelete: false})}}
+            okButtonText="Delete"
+          ></PromptBox>
         }
       </div>
     )
